+++ table_init_z.h
@@ -11,7 +11,7 @@
 
 namespace cvc {
 
-inline double _Complex * init_1level_ztable ( int const N0 ) {
+inline double _Complex * init_1level_ztable ( size_t const N0 ) {
   return( ( double _Complex *) calloc ( N0 , sizeof( double _Complex ) ) );
 }  // end of init_1level_ztable
 
@@ -27,7 +27,7 @@ inline void fini_1level_ztable ( double _Complex **s  ) {
 /************************************************************************************/
 /************************************************************************************/
 
-inline double _Complex ** init_2level_ztable (int const N0, int const N1 ) {
+inline double _Complex ** init_2level_ztable (size_t const N0, size_t const N1 ) {
   double _Complex * s__ = NULL;
   s__ = init_1level_ztable ( N0*N1);
   if ( s__ == NULL ) return( NULL );
@@ -35,7 +35,7 @@ inline double _Complex ** init_2level_ztable (int const N0, int const N1 ) {
   double _Complex ** s_ = ( double _Complex **) malloc( N0 * sizeof( double _Complex *) );
   if ( s_ == NULL ) return ( NULL );
 
-  for ( int i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
+  for ( size_t i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
   return( s_ );
 }  // end of init_2level_ztable
 
@@ -56,7 +56,7 @@ inline void fini_2level_ztable ( double _Complex *** s  ) {
 /************************************************************************************/
 
 
-inline double _Complex *** init_3level_ztable (int const N0, int const N1, int const N2 ) {
+inline double _Complex *** init_3level_ztable (size_t const N0, size_t const N1, size_t const N2 ) {
   double _Complex ** s__ = NULL;
   s__ = init_2level_ztable ( N0*N1, N2);
   if ( s__ == NULL ) return( NULL );
@@ -64,7 +64,7 @@ inline double _Complex *** init_3level_ztable (int const N0, int const N1, int c
   double _Complex *** s_ = ( double _Complex ***) malloc( N0 * sizeof( double _Complex **) );
   if ( s_ == NULL ) return ( NULL );
 
-  for ( int i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
+  for ( size_t i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
   return( s_ );
 }  // end of init_3level_ztable
 
@@ -85,7 +85,7 @@ inline void fini_3level_ztable ( double _Complex **** s  ) {
 /************************************************************************************/
 
 
-inline double _Complex **** init_4level_ztable (int const N0, int const N1, int const N2, int const N3 ) {
+inline double _Complex **** init_4level_ztable (size_t const N0, size_t const N1, size_t const N2, size_t const N3 ) {
   double _Complex *** s__ = NULL;
   s__ = init_3level_ztable ( N0*N1, N2, N3);
   if ( s__ == NULL ) return( NULL );
@@ -93,7 +93,7 @@ inline double _Complex **** init_4level_ztable (int const N0, int const N1, int 
   double _Complex **** s_ = ( double _Complex ****) malloc( N0 * sizeof( double _Complex ***) );
   if ( s_ == NULL ) return ( NULL );
 
-  for ( int i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
+  for ( size_t i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
   return( s_ );
 }  // end of init_4level_ztable
 
@@ -114,7 +114,7 @@ inline void fini_4level_ztable ( double _Complex ***** s  ) {
 /************************************************************************************/
 
 
-inline double _Complex ***** init_5level_ztable (int const N0, int const N1, int const N2, int const N3, int const N4 ) {
+inline double _Complex ***** init_5level_ztable (size_t const N0, size_t const N1, size_t const N2, size_t const N3, size_t const N4 ) {
   double _Complex **** s__ = NULL;
   s__ = init_4level_ztable ( N0*N1, N2, N3, N4);
   if ( s__ == NULL ) return( NULL );
@@ -122,7 +122,7 @@ inline double _Complex ***** init_5level_ztable (int const N0, int const N1, int
   double _Complex ***** s_ = ( double _Complex *****) malloc( N0 * sizeof( double _Complex ****) );
   if ( s_ == NULL ) return ( NULL );
 
-  for ( int i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
+  for ( size_t i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
   return( s_ );
 }  // end of init_5level_ztable
 
@@ -143,7 +143,7 @@ inline void fini_5level_ztable ( double _Complex ****** s  ) {
 /************************************************************************************/
 
 
-inline double _Complex ****** init_6level_ztable (int const N0, int const N1, int const N2, int const N3, int const N4, int const N5 ) {
+inline double _Complex ****** init_6level_ztable (size_t const N0, size_t const N1, size_t const N2, size_t const N3, size_t const N4, size_t const N5 ) {
   double _Complex ***** s__ = NULL;
   s__ = init_5level_ztable ( N0*N1, N2, N3, N4, N5);
   if ( s__ == NULL ) return( NULL );
@@ -151,7 +151,7 @@ inline double _Complex ****** init_6level_ztable (int const N0, int const N1, in
   double _Complex ****** s_ = ( double _Complex ******) malloc( N0 * sizeof( double _Complex *****) );
   if ( s_ == NULL ) return ( NULL );
 
-  for ( int i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
+  for ( size_t i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
   return( s_ );
 }  // end of init_6level_ztable
 
@@ -172,7 +172,7 @@ inline void fini_6level_ztable ( double _Complex ******* s  ) {
 /************************************************************************************/
 
 
-inline double _Complex ******* init_7level_ztable (int const N0, int const N1, int const N2, int const N3, int const N4, int const N5, int const N6 ) {
+inline double _Complex ******* init_7level_ztable (size_t const N0, size_t const N1, size_t const N2, size_t const N3, size_t const N4, size_t const N5, size_t const N6 ) {
   double _Complex ****** s__ = NULL;
   s__ = init_6level_ztable ( N0*N1, N2, N3, N4, N5, N6);
   if ( s__ == NULL ) return( NULL );
@@ -180,7 +180,7 @@ inline double _Complex ******* init_7level_ztable (int const N0, int const N1, i
   double _Complex ******* s_ = ( double _Complex *******) malloc( N0 * sizeof( double _Complex ******) );
   if ( s_ == NULL ) return ( NULL );
 
-  for ( int i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
+  for ( size_t i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
   return( s_ );
 }  // end of init_7level_ztable
 
@@ -201,7 +201,7 @@ inline void fini_7level_ztable ( double _Complex ******** s  ) {
 /************************************************************************************/
 
 
-inline double _Complex ******** init_8level_ztable (int const N0, int const N1, int const N2, int const N3, int const N4, int const N5, int const N6, int const N7 ) {
+inline double _Complex ******** init_8level_ztable (size_t const N0, size_t const N1, size_t const N2, size_t const N3, size_t const N4, size_t const N5, size_t const N6, size_t const N7 ) {
   double _Complex ******* s__ = NULL;
   s__ = init_7level_ztable ( N0*N1, N2, N3, N4, N5, N6, N7);
   if ( s__ == NULL ) return( NULL );
@@ -209,7 +209,7 @@ inline double _Complex ******** init_8level_ztable (int const N0, int const N1, 
   double _Complex ******** s_ = ( double _Complex ********) malloc( N0 * sizeof( double _Complex *******) );
   if ( s_ == NULL ) return ( NULL );
 
-  for ( int i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
+  for ( size_t i = 0; i < N0; i++ ) s_[i] = s__ + i * N1;
   return( s_ );
 }  // end of init_8level_ztable
